IT Companies Agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "devpulse-ai-backend"
version = "0.1.0"
description = "Backend architecture scaffold for DevPulse AI"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
  "asyncpg>=0.29.0",
  "pyjwt>=2.12.1",
  "google-generativeai>=0.5.0",
  "chromadb>=0.4.24",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

